a shit ton of changes
so this is supposed to be used for handling events and incoming messages, just for making it easy to work with.

The idea next is to actually start sending information between the nodes and start improving how it works.

outside of that it doesnt seems like if it needs a lot of shit just to work, now i think that is time to actually adapt it to the new idea